mutt: Move config to ~/.mutt and use multiple muttrc files
This moves existing config under ~/.mutt to .config/mutt.

It also moves .muttrc to .config/mutt. In my updated environment I
stopped using a .muttrc but require mutt to be invoked with -F to load
~/.muttrc-mboxorg or ~/.muttrc-netscout for current employer. I also
have a gmail account I use to test IMAP features out at times as
software is written with GMail in mind now days so if I am making
changes I like to use Google for the real test.